requirements.txt
discord.py>=2.4
aiohttp>=3.9
beautifulsoup4>=4.12
lxml>=5.0
feedparser>=6.0
aiosqlite>=0.19
python-dotenv>=1.0
